% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dima.template.R
\name{dima.template}
\alias{dima.template}
\alias{build.DIMA.template}
\title{Build DIMA Template Files}
\usage{
build.DIMA.template(data, template.file, method, out.file)
}
\arguments{
\item{data}{The data output from xlsx2R.}

\item{template.file}{An Excel of a blank DIMA template}

\item{method}{The file template method. Options are "LPI", "Richness", "Gap".}

\item{out.file}{The file path for the output}
}
\value{
A \code {tbl} of files formated for a DIMA template.
}
\description{
Build a file in the format needed for DIMA
}
